requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
dnspython==2.1.0
filelock==3.0.12
idna==3.2
py3-validate-email==1.0.2
pytz==2021.1
sqlparse==0.4.2
